refactor(runner): move stageFlowFiles into its own module
prepareRunDir.ts was one line under the 150-line cap, leaving no room in the staging code.
